Utility libraries and Arduino sketch for greenhouse probes
The libraries are mostly a set of simple utilities that ships off as much as I can to C++, since .ino is a pretty limited language. The sketch is set up to be able to log data to a server somewhere, except for the part where it's doing none of that.

Most of the reads/writes are done using low-level bit sets for the Yun specifically. I have no intention of ever generalising this code to a different device. Most of the sensor reads were happening in the order of like 20ms, IIRC? so pretty fast.

Ideally wasNack belongs in I2C.h but I can't really be bothered making an upstream PR for this.